requirements.txt
# Core de la aplicación y IA
streamlit
langchain
langchain-google-genai
google-generativeai
python-dotenv
pydantic

# Interacción por Voz
SpeechRecognition>=3.10.1
elevenlabs
playsound==1.2.2

# Manipulación de Archivos y Documentos
cloudconvert
python-docx
PyPDF2
Pillow
pdf2docx
docx2pdf

# Comunicación y APIs
requests>=2.31.0
python-multipart

# gRPC para Microservicio Mangle
grpcio
grpcio-tools
